+++ common/src/main/java/org/apache/celeborn/common/network/server/memory/BufferRecycler.java
@@ -15,16 +15,30 @@
  * limitations under the License.
  */
 
-package org.apache.celeborn.common.memory;
+package org.apache.celeborn.common.network.server.memory;
+
+import java.util.function.Consumer;
 
 import io.netty.buffer.ByteBuf;
 
-/** manager bytebuf lifecycle */
-public interface Recycler {
+public class BufferRecycler implements Recycler {
 
-  /** recycle bytebuf to buffer pool */
-  void recycle(ByteBuf byteBuf);
+  private MemoryManager memoryManager;
 
-  /** release the bytebuf */
-  void release(ByteBuf byteBuf);
+  private Consumer<ByteBuf> recycleConsumer;
+
+  public BufferRecycler(MemoryManager memoryManager, Consumer<ByteBuf> recycleConsumer) {
+    this.memoryManager = memoryManager;
+    this.recycleConsumer = recycleConsumer;
+  }
+
+  @Override
+  public void recycle(ByteBuf byteBuf) {
+    recycleConsumer.accept(byteBuf);
+  }
+
+  @Override
+  public void release(ByteBuf byteBuf) {
+    memoryManager.recycleReadBuffer(byteBuf);
+  }
 }
