+++ common/src/main/java/org/apache/celeborn/common/network/server/memory/WrappedDataBuffer.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.celeborn.common.memory;
+package org.apache.celeborn.common.network.server.memory;
 
 import io.netty.buffer.ByteBuf;
