+++ worker/src/test/scala/org/apache/celeborn/service/deploy/worker/fetch/credit/BufferStreamManagerSuiteJ.java
@@ -36,7 +36,7 @@ import org.slf4j.LoggerFactory;
 
 import org.apache.celeborn.common.identity.UserIdentifier;
 import org.apache.celeborn.common.meta.FileInfo;
-import org.apache.celeborn.common.memory.MemoryManager;
+import org.apache.celeborn.common.network.server.memory.MemoryManager;
 import org.apache.celeborn.common.util.JavaUtils;
 import org.apache.celeborn.common.util.Utils;
